+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix bug where editing resets progress.
 - Tap word hints area to show next word.
 - Hide hint box when showing answer.
+- Reset due dates for entire collection.
 
 ## 1.3.0 - July 26, 2023